implement a more friendly derive macro for State
State can be derived on Unit structs (as Type::Unit)
State can be derived on Tuple structs (as Type::List)

`#[state_ignore]` still exists but:
`#[anathema(ignore)]` is a more 'unified' form of it

and I added:
`#[anathema(rename = "name")]` which can be used
to rename a field from the `aml` side of things

Note: the attributes cannot be currently used in
unit/tuple structs.

Example:

```rust
use anathema::{State, state::Value}; // TODO prelude names for this

struct AppState {
    #[anathema(ignore)]
    inner: bool

    #[anathema(rename = "text")]
    data: Value<String>
}
```

This'll allow the state's data to be referenced as
`.text` from the template, but stay as 'data' on
the rust side.